+++ model/result.go
@@ -4,5 +4,5 @@ type Result struct {
 	Code  int    `json:"code" default:"0"`
 	Msg   string `json:"msg"`
 	Data  any    `json:"data"`
-	Count int    `json:"count" default:"0"`
+	Count int64  `json:"count" default:"0"`
 }
